patch u8 combo to u16

+++ contracts/src/elements/bonuses/hammer.cairo
@@ -38,7 +38,7 @@ impl BonusImpl of BonusTrait {
     }
 
     #[inline(always)]
-    fn get_count(score: u32, combo_count: u8, max_combo: u8) -> u8 {
+    fn get_count(score: u32, combo_count: u16, max_combo: u8) -> u8 {
         if score >= 120 {
             return 3;
         }
